divide align from element_commands
